Add Joke and Text cogs with various text manipulation commands
- Implemented JokeCog with a command to fetch random jokes from an API.
- Added TextCog featuring multiple text transformation commands including catify, emojify, reverse, drunkify, formalify, and more.
- Updated TranslateCog to use async translation methods for improved performance.
